services/joblook-backend/pyproject.toml
[project]
name = "joblook-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "jinja2>=3.1",
    "python-multipart>=0.0.12",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg[binary]>=3.2",
    "pydantic[email]>=2.9",
    "pydantic-settings>=2.6",
    "argon2-cffi>=23.1",
    "cryptography>=43.0",
    "itsdangerous>=2.2",
    "anthropic>=0.40",
    "pypdf>=5.1",
    "python-docx>=1.1",
    "authlib>=1.3",
    "httpx>=0.27",
]

[tool.uv]
package = false

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
